GHSA-r74q-gxcg-73hx workaround for simple_form
Workarounds
The issue only happens with pages that build forms based on user-provided input. If your application doesn't do that, you're not affected.
A workaround is to explicitly pass which type you want for an input since the issue lies on Simple Form's automatically discovery of input types. This can be done using the as option, like the following:

<%= form.input :avatar, as: :file %>